add duplicate add prevention for history

+++ src/core/controllers/websites/remove/remove.ts
@@ -42,10 +42,10 @@ export const removeWebsite = async ({ userId, url, deleteMany = false }) => {
   if (typeof userId !== "undefined") {
     const deleteQuery = { domain: siteExist?.domain, userId };
 
-    const [_, historyCollection] = await HistoryController().getHistoryItem(
-      deleteQuery,
-      true
-    );
+    const [
+      history,
+      historyCollection,
+    ] = await HistoryController().getHistoryItem(deleteQuery, true);
 
     await scriptsCollection.deleteMany(deleteQuery);
     await analyticsCollection.deleteMany(deleteQuery);
@@ -53,13 +53,16 @@ export const removeWebsite = async ({ userId, url, deleteMany = false }) => {
     await issuesCollection.deleteMany(deleteQuery);
     await collection.findOneAndDelete(deleteQuery);
 
-    if ((await historyCollection.countDocuments({ userId })) >= 20) {
-      await historyCollection.deleteOne({ userId });
+    // PREVENT DUPLICATE ITEMS IN HISTORY
+    if (!history) {
+      if ((await historyCollection.countDocuments({ userId })) >= 20) {
+        await historyCollection.deleteOne({ userId });
+      }
+      await historyCollection.insertOne({
+        ...siteExist,
+        deletedDate: new Date(),
+      });
     }
-    await historyCollection.insertOne({
-      ...siteExist,
-      deletedDate: new Date(),
-    });
 
     return { website: siteExist, code: 200, success: true, message: SUCCESS };
   }
